requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings>=2.0.0

# Machine Learning
torch>=2.2.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Template Engine
jinja2==3.1.2

# Database (SQLite for now)
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
